Fixed errors in test cases and moved DummyTermination to gym_solo.testing
